requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2024.10
cryptography==44.0.1
cssselect2==0.7.0
Django==5.1.6
django-crispy-forms==2.3
django-filter==25.1
django-jalali==7.3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.56.0
html5lib==1.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
lxml==5.3.1
numpy==2.2.3
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.9.0
pypdf==5.3.0
pyphen==0.17.2
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.1
tzlocal==5.3
Unidecode==1.3.8
uritools==4.0.3
urllib3==2.3.0
weasyprint==64.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
